Update LexerUtils.scala
Optimized the logic of LexerUtils.scala for blank line judgment.
When the last word on the previous line of the cursor is ";", the complex judgment logic that follows will no longer be executed.

Update LexerUtilsTest

Added unit tests for corresponding optimizations。
And modified an error in the loginfo of the AutoSuggestContext.scala file